refactor: move plugin to its own crate
