+++ devtools/Cargo.toml
@@ -29,6 +29,7 @@ bytes = "1.5.0"
 ringbuf = "0.4.0-rc.2"
 colored = "2.0"
 async-stream = "0.3.5"
+http = "0.2"
 
 [dev-dependencies]
 tauri = { workspace = true, features = ["test"] }
