+++ devtools/src/layer.rs
@@ -29,7 +29,7 @@ thread_local! {
 /// A tracing layer that forwards all events to the aggregator
 /// This is intentionally kept as simple as possible to avoid any performance overhead on
 /// the application thread. All the heavy lifting is done in the aggregator.
-pub(crate) struct Layer {
+pub struct Layer {
     shared: Arc<Shared>,
     tx: mpsc::Sender<Event>,
 }
